refactor: replace CKEditor build dir with build file only
