+++ components/admin/TextEditor/custom-build/ckeditor.js
