+++ lib/hooks/useCKEditor.js
@@ -9,7 +9,7 @@ const useCKEditor = () => {
     editorRef.current = {
       /* eslint-disable node/global-require */
       CKEditor: require( '@ckeditor/ckeditor5-react' ).CKEditor,
-      Editor: require( 'ckeditor5-custom-build/build/ckeditor' ),
+      Editor: require( '../../components/admin/TextEditor/custom-build/ckeditor' ),
     };
 
     setIsEditorLoaded( true );
